src/debugAdapter: make filepaths in errors absolute
Errors from delve often contain relative paths, which may not be
relative to the workspace directory resulting in bad links. Try to
make the paths absolute before displaying to the user.

Fixes golang/vscode-go#456

Reviewed-on: https://go-review.googlesource.com/c/vscode-go/+/252978
Run-TryBot: Suzy Mueller <suzmue@golang.org>
TryBot-Result: kokoro <noreply+kokoro@google.com>
